CI issues fix
